move verge to own directory
