+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/fagsak/Fagsak.kt
@@ -2,7 +2,7 @@ package no.nav.familie.ba.sak.kjerne.fagsak
 
 import no.nav.familie.ba.sak.common.BaseEntitet
 import no.nav.familie.ba.sak.kjerne.institusjon.Institusjon
-import no.nav.familie.ba.sak.kjerne.institusjon.Verge
+import no.nav.familie.ba.sak.kjerne.verge.Verge
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import java.util.Objects
 import javax.persistence.Column
